separates crud features into traits; fixes #128; fixes #95; fixes #134
