+++ src/app/Http/Controllers/CrudController.php
@@ -2,19 +2,25 @@
 
 namespace Backpack\CRUD\app\Http\Controllers;
 
+use Illuminate\Foundation\Bus\DispatchesJobs;
+use Illuminate\Foundation\Validation\ValidatesRequests;
+use Illuminate\Routing\Controller as BaseController;
+use Illuminate\Support\Facades\Form as Form;
+
 use Backpack\CRUD\app\Http\Requests\CrudRequest as StoreRequest;
 use Backpack\CRUD\app\Http\Requests\CrudRequest as UpdateRequest;
 use Backpack\CRUD\CrudPanel;
-use Illuminate\Foundation\Bus\DispatchesJobs;
-use Illuminate\Foundation\Validation\ValidatesRequests;
-// VALIDATION: change the requests to match your own file names if you need form validation
-use Illuminate\Routing\Controller as BaseController;
-use Illuminate\Support\Facades\Form as Form;
-use LiveControl\EloquentDataTable\DataTable;
+
+// CRUD Traits for non-core features
+use Backpack\CRUD\app\Http\Controllers\CrudFeatures\AjaxTable;
+use Backpack\CRUD\app\Http\Controllers\CrudFeatures\Reorder;
+use Backpack\CRUD\app\Http\Controllers\CrudFeatures\Revisions;
+use Backpack\CRUD\app\Http\Controllers\CrudFeatures\ShowDetailsRow;
 
 class CrudController extends BaseController
 {
     use DispatchesJobs, ValidatesRequests;
+    use AjaxTable, Reorder, Revisions, ShowDetailsRow;
 
     public $data = [];
     public $crud;
@@ -22,7 +28,20 @@ class CrudController extends BaseController
     public function __construct()
     {
         $this->crud = new CrudPanel();
+
+        // call the setup function inside this closure to also have the request there
+        // this way, developers can use things stored in session (auth variables, etc)
+        $this->middleware(function ($request, $next) {
+            $this->setup();
+            return $next($request);
+        });
     }
+
+    /**
+     * Allow developers to set their configuration options for a CrudPanel.
+     */
+    public function setup() {}
+
 
     /**
      * Display all rows in the database for this entity.
@@ -88,7 +107,7 @@ class CrudController extends BaseController
         }
 
         // insert item in the db
-        $item = $this->crud->create($request->except(['redirect_after_save', 'password', '_token']));
+        $item = $this->crud->create($request->except(['redirect_after_save', '_token']));
 
         // show a success message
         \Alert::success(trans('backpack::crud.insert_success'))->flash();
@@ -180,61 +199,6 @@ class CrudController extends BaseController
     }
 
     /**
-     * Display the revisions for specified resource.
-     *
-     * @param int $id
-     *
-     * @return Response
-     */
-    public function listRevisions($id)
-    {
-        $this->crud->hasAccessOrFail('revisions');
-
-        // get the info for that entry
-        $this->data['entry'] = $this->crud->getEntry($id);
-        $this->data['crud'] = $this->crud;
-        $this->data['title'] = ucfirst($this->crud->entity_name).' '.trans('backpack::crud.revisions');
-        $this->data['id'] = $id;
-        $this->data['revisions'] = $this->crud->listRevisions($id);
-
-        // load the view from /resources/views/vendor/backpack/crud/ if it exists, otherwise load the one in the package
-        return view('crud::revisions', $this->data);
-    }
-
-    /**
-     * Restore a specific revision for the specified resource.
-     *
-     * Used via AJAX in the revisions view
-     *
-     * @param int $id
-     *
-     * @return JSON Response containing the new revision that was created from the update
-     * @return HTTP 500 if the request did not contain the revision ID
-     */
-    public function restoreRevision($id)
-    {
-        $this->crud->hasAccessOrFail('revisions');
-
-        // @TODO: Currently the route already contains the revision ID, so passing it as a POST param
-        // is somewhat superfluous.. however if we are POSTing, it makes sense to actually have data to post.
-        // Perhaps the route shoud be better named to reflect this (e.g. just /model/{id}/revisions) (??)
-        $revisionId = \Request::input('revision_id', false);
-        if (! $revisionId) {
-            abort(500, 'Can\'t restore revision without revision_id');
-        } else {
-            $this->crud->restoreRevision($id, $revisionId); // do the update
-
-            $this->data['entry'] = $this->crud->getEntry($id);
-            $this->data['crud'] = $this->crud;
-            $this->data['revisions'] = $this->crud->listRevisions($id); // Reload revisions as they have changed
-
-            // Rebuild the revision timeline HTML and return it to the AJAX call
-            // @TODO: Return only the latest revision to save bandwidth - 15/9/16 @se1exin
-            return view('crud::inc.revision_timeline', $this->data);
-        }
-    }
-
-    /**
      * Remove the specified resource from storage.
      *
      * @param int $id
@@ -246,120 +210,5 @@ class CrudController extends BaseController
         $this->crud->hasAccessOrFail('delete');
 
         return $this->crud->delete($id);
-    }
-
-    /**
-     *  Reorder the items in the database using the Nested Set pattern.
-     *
-     *	Database columns needed: id, parent_id, lft, rgt, depth, name/title
-     *
-     *  @return Response
-     */
-    public function reorder()
-    {
-        $this->crud->hasAccessOrFail('reorder');
-
-        if (! $this->crud->isReorderEnabled()) {
-            abort(403, 'Reorder is disabled.');
-        }
-
-        // get all results for that entity
-        $this->data['entries'] = $this->crud->getEntries();
-        $this->data['crud'] = $this->crud;
-        $this->data['title'] = trans('backpack::crud.reorder').' '.$this->crud->entity_name;
-
-        // load the view from /resources/views/vendor/backpack/crud/ if it exists, otherwise load the one in the package
-        return view('crud::reorder', $this->data);
-    }
-
-    /**
-     * Save the new order, using the Nested Set pattern.
-     *
-     * Database columns needed: id, parent_id, lft, rgt, depth, name/title
-     *
-     * @return
-     */
-    public function saveReorder()
-    {
-        $this->crud->hasAccessOrFail('reorder');
-
-        $all_entries = \Request::input('tree');
-
-        if (count($all_entries)) {
-            $count = $this->crud->updateTreeOrder($all_entries);
-        } else {
-            return false;
-        }
-
-        return 'success for '.$count.' items';
-    }
-
-    /**
-     * Used with AJAX in the list view (datatables) to show extra information about that row that didn't fit in the table.
-     * It defaults to showing some dummy text.
-     *
-     * It's enabled by:
-     * - setting: $crud->details_row = true;
-     * - adding the details route for the entity; ex: Route::get('page/{id}/details', 'PageCrudController@showDetailsRow');
-     * - adding a view with the following name to change what the row actually contains: app/resources/views/vendor/backpack/crud/details_row.blade.php
-     */
-    public function showDetailsRow($id)
-    {
-        $this->crud->hasAccessOrFail('details_row');
-
-        $this->data['entry'] = $this->crud->getEntry($id);
-        $this->data['crud'] = $this->crud;
-
-        // load the view from /resources/views/vendor/backpack/crud/ if it exists, otherwise load the one in the package
-        return view('crud::details_row', $this->data);
-    }
-
-    /**
-     * Respond with the JSON of one or more rows, depending on the POST parameters.
-     * @return JSON Array of cells in HTML form.
-     */
-    public function search()
-    {
-        $this->crud->hasAccessOrFail('list');
-
-        // crate an array with the names of the searchable columns
-        $columns = collect($this->crud->columns)
-                    ->reject(function ($column, $key) {
-                        // the select_multiple columns are not searchable
-                        return isset($column['type']) && $column['type'] == 'select_multiple';
-                    })
-                    ->pluck('name')
-                    // add the primary key, otherwise the buttons won't work
-                    ->merge($this->crud->model->getKeyName())
-                    ->toArray();
-
-        // structure the response in a DataTable-friendly way
-        $dataTable = new DataTable($this->crud->query, $columns);
-
-        // make the datatable use the column types instead of just echoing the text
-        $dataTable->setFormatRowFunction(function ($entry) {
-            // get the actual HTML for each row's cell
-            $row_items = $this->crud->getRowViews($entry, $this->crud);
-
-            // add the buttons as the last column
-            if ($this->crud->buttons->where('stack', 'line')->count()) {
-                $row_items[] = \View::make('crud::inc.button_stack', ['stack' => 'line'])
-                                ->with('crud', $this->crud)
-                                ->with('entry', $entry)
-                                ->render();
-            }
-
-            // add the details_row buttons as the first column
-            if ($this->crud->details_row) {
-                array_unshift($row_items, \View::make('crud::columns.details_row_button')
-                                ->with('crud', $this->crud)
-                                ->with('entry', $entry)
-                                ->render());
-            }
-
-            return $row_items;
-        });
-
-        return $dataTable->make();
     }
 }
